requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas
numpy
PyYAML~=5.3.1
setuptools~=51.3.3
scikit-learn
plotly-express~=0.4.0
hyperopt